Began to work on the filtering with the input
