+++ src/pages/Home.tsx
@@ -1,10 +1,10 @@
-import { FilterCourse } from "../components/FilterCourse"
+import { CourseAdd } from "../components/CourseAdd"
 
 export const Home = () => {
     return(
         <div>
             <h1>Homepage</h1>
-            <FilterCourse />
+            <CourseAdd />
         </div>
     )
 }
